% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_caching.R
\name{tw_check_cache}
\alias{tw_check_cache}
\title{Enable caching for the current session}
\usage{
tw_check_cache(cache = NULL)
}
\value{
Nothing, used for its side effects.
}
\description{
Enable caching for the current session
}
\examples{
\dontrun{
tw_check_cache()
}
}
