% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_caching.R
\name{tw_check_cache_folder}
\alias{tw_check_cache_folder}
\title{Checks if cache folder exists, if not returns an informative message}
\usage{
tw_check_cache_folder()
}
\value{
If the cache folder exists, returns TRUE. Otherwise throws an error.
}
\description{
Checks if cache folder exists, if not returns an informative message
}
\examples{

# If cache folder does not exist, it throws an error
tryCatch(tw_check_cache_folder(),
  error = function(e) {
    return(e)
  }
)

# Create cache folder
tw_set_cache_folder(path = fs::path(
  tempdir(),
  "tw_cache_folder"
))
tw_create_cache_folder(ask = FALSE)

tw_check_cache_folder()
}
