% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_caching.R
\name{tw_create_cache_folder}
\alias{tw_create_cache_folder}
\title{Creates the cache folder where \code{streetnames} caches data.}
\usage{
tw_create_cache_folder(ask = TRUE)
}
\arguments{
\item{ask}{Logical, defaults to TRUE. If FALSE, and cache folder does not exist, it just creates it without asking (useful for interactive sessions).}
}
\value{
Nothing, used for its side effects.
}
\description{
Creates the cache folder where \code{streetnames} caches data.
}
\examples{
\dontrun{
tw_create_cache_folder()
}
}
