% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_caching.R
\name{tw_disable_cache}
\alias{tw_disable_cache}
\title{Disable caching for the current session}
\usage{
tw_disable_cache()
}
\value{
Nothing, used for its side effects.
}
\description{
Disable caching for the current session
}
\examples{
\donttest{
if (interactive()) {
  tw_disable_cache()
}
}
}
