% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_caching.R
\name{tw_disconnect_from_cache}
\alias{tw_disconnect_from_cache}
\title{Ensure that connection to cache is disconnected consistently}
\usage{
tw_disconnect_from_cache(
  cache = NULL,
  cache_connection = NULL,
  disconnect_db = TRUE,
  language = tidywikidatar::tw_get_language()
)
}
\arguments{
\item{cache}{Defaults to NULL. If given, it should be given either TRUE or FALSE. Typically set with \code{tw_enable_cache()} or \code{tw_disable_cache()}.}

\item{cache_connection}{Defaults to NULL. If NULL, and caching is enabled, \code{tidywikidatar} will use a local sqlite database. A custom connection to other databases can be given (see vignette \code{caching} for details).}

\item{disconnect_db}{Defaults to TRUE. If FALSE, leaves the connection to cache open.}
}
\value{
Nothing, used for its side effects.
}
\description{
Ensure that connection to cache is disconnected consistently
}
\examples{

tw_get(
  id = c("Q180099"),
  language = "en"
)
tw_disconnect_from_cache()

}
