% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tw_empty_image_metadata}
\alias{tw_empty_image_metadata}
\title{A zero-rows tibble used internally when \code{tw_get_image_metadata()} would not return any value.}
\format{
A data frame with 0 rows and 19 columns
}
\usage{
tw_empty_image_metadata
}
\description{
A zero-rows tibble used internally when \code{tw_get_image_metadata()} would not return any value.
}
\keyword{datasets}
