% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tw_empty_search}
\alias{tw_empty_search}
\title{A zero-rows tibble used internally when \code{tw_search()} would not return any value.}
\format{
A data frame with 0 rows and 4 columns
}
\usage{
tw_empty_search
}
\description{
A zero-rows tibble used internally when \code{tw_search()} would not return any value.
}
\keyword{datasets}
