% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tw_empty_wikipedia_page}
\alias{tw_empty_wikipedia_page}
\title{A zero-rows tibble used internally when \code{tw_get_qid_of_wikipedia_page()} or \code{tw_get_links_from_wikipedia_page()} would not return any value.}
\format{
A data frame with 0 rows and 6 columns
}
\usage{
tw_empty_wikipedia_page
}
\description{
A zero-rows tibble used internally when \code{tw_get_qid_of_wikipedia_page()} or \code{tw_get_links_from_wikipedia_page()} would not return any value.
}
\keyword{datasets}
