% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tw_empty_wikipedia_page_sections}
\alias{tw_empty_wikipedia_page_sections}
\title{A zero-rows tibble used internally when \code{tw_get_wikipedia_page_sections()} would not return any value.}
\format{
A data frame with 0 rows and 8 columns
}
\usage{
tw_empty_wikipedia_page_sections
}
\description{
A zero-rows tibble used internally when \code{tw_get_wikipedia_page_sections()} would not return any value.
}
\keyword{datasets}
