% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_caching.R
\name{tw_enable_cache}
\alias{tw_enable_cache}
\title{Enable caching for the current session}
\usage{
tw_enable_cache(SQLite = TRUE)
}
\arguments{
\item{SQLite}{Logical, defaults to TRUE. Set to FALSE to use custom database options. See \code{tw_set_cache_db()} for details.}
}
\value{
Nothing, used for its side effects.
}
\description{
Enable caching for the current session
}
\examples{
\donttest{
if (interactive()) {
  tw_enable_cache()
}
}
}
