% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_filter.R
\name{tw_filter}
\alias{tw_filter}
\title{Filter search result and keep only people}
\usage{
tw_filter(
  search,
  p,
  q,
  stop_at_first = TRUE,
  language = "en",
  limit = 10,
  wait = 1,
  cache = TRUE
)
}
\arguments{
\item{search}{A data frame generated by \code{tw_search()}, or a search query. If a data frame is given, language and limits are ignore.}

\item{p}{A character vector of length 1, a property. Must always start with the capital letter "P", e.g. "P31" for "instance of".}

\item{q}{A character vector of length 1, a wikidata id. Must always start with the capital letter "Q", e.g. "Q5" for "human being".}

\item{language}{Language to be used for the search.}

\item{limit}{Maximum numbers of responses to be given.}

\item{wait}{In seconds, defaults to 1. Time to wait between queries to Wikidata. If data are cached locally, wait time is not applied.}

\item{cache}{Logical, defaults to TRUE.}
}
\value{
A data frame with three columns, \code{id}, \code{label}, and \code{description}, filtered by the above criteria.
}
\description{
Filter search result and keep only people
}
\examples{

\dontrun{
tw_search("Antonio Vivaldi") \%>\%
  tw_filter()
}

}
