% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_caching.R
\name{tw_get_cache_db}
\alias{tw_get_cache_db}
\title{Get database connection settings from the environment}
\usage{
tw_get_cache_db()
}
\value{
A list with all database parameters as stored in environment variables.
}
\description{
Typically set with \code{tw_set_cache_db()}
}
\examples{

tw_get_cache_db()
}
