% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_get_cached_item.R
\name{tw_get_cache_file}
\alias{tw_get_cache_file}
\title{Gets location of cache file}
\usage{
tw_get_cache_file(
  extension = "sqlite",
  language = tidywikidatar::tw_get_language()
)
}
\arguments{
\item{extension}{Defaults to \code{sqlite}. Extension of the database cache file.}

\item{language}{Defaults to language set with \code{\link[=tw_set_language]{tw_set_language()}}; if not
set, "en". Use "all_available" to keep all languages. For available
language values, see
\href{https://www.wikidata.org/wiki/Help:Wikimedia_language_codes/lists/all}{the
dedicated Wikimedia page}.}
}
\value{
A character vector of length one with location of item cache file.
}
\description{
Gets location of cache file
}
\examples{

tw_set_cache_folder(path = tempdir())
sqlite_cache_file_location <- tw_get_cache_file() # outputs location of cache file
}
