% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_get.R
\name{tw_get_image}
\alias{tw_get_image}
\title{Get Wikidata image}
\usage{
tw_get_image(id, cache = NULL, overwrite_cache = FALSE, format = "filename")
}
\arguments{
\item{id}{A characther vector of length 1, must start with Q, e.g. "Q254" for Wolfgang Amadeus Mozart.}

\item{cache}{Defaults to NULL. If given, it should be given either TRUE or FALSE. Typically set with \code{tw_enable_cache()} or \code{tw_disable_cache()}.}

\item{overwrite_cache}{Logical, defaults to FALSE. If TRUE, it overwrites the table in the local sqlite database. Useful if the original Wikidata object has been updated.}

\item{format}{A charachter vector, defaults to 'filename". If set to 'commons', outputs the link to the Wikimedia Commons page. If set to "embed", outputs a link that can be used to embed.}
}
\value{
A charachter vector, corresponding to reference to the image in the requested format.
}
\description{
Get Wikidata image
}
\examples{
\dontrun{
tw_get_image(id = "Q228822")
}

}
