% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_get.R
\name{tw_get_label}
\alias{tw_get_label}
\title{Get Wikidata label in given language}
\usage{
tw_get_label(id, language = "en", cache = TRUE, overwrite_cache = FALSE)
}
\arguments{
\item{id}{A characther vector, must start with Q, e.g. "Q254" for Wolfgang Amadeus Mozart}

\item{language}{A character vector of length one, defaults to "en", must correspond to a two-letter language code.}

\item{cache}{Logical, defaults to TRUE. If TRUE, it stores all retrieved data in a local sqlite database.}

\item{overwrite_cache}{Logical, defaults to FALSE. If TRUE, it overwrites the table in the local sqlite database. Useful if the original Wikidata object has been updated.}
}
\value{
A charachter vector of length 1, with the Wikidata label in the requested languae.
}
\description{
Get Wikidata label in given language
}
\examples{


}
