% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_get_id_from_wikipedia.R
\name{tw_get_qid_of_wikipedia_page_single}
\alias{tw_get_qid_of_wikipedia_page_single}
\title{Gets the Wikidata id of a Wikipedia page}
\usage{
tw_get_qid_of_wikipedia_page_single(
  url = NULL,
  title = NULL,
  language = tidywikidatar::tw_get_language(),
  cache = NULL,
  overwrite_cache = FALSE,
  cache_connection = NULL,
  disconnect_db = TRUE
)
}
\arguments{
\item{url}{Full URL to a Wikipedia page. If given, title and language can be left empty.}

\item{title}{Title of a Wikipedia page or final parts of its url. If given, url can be left empty, but language must be provided.}

\item{language}{Two-letter language code used to define the Wikipedia version to use. Defaults to language set with \code{tw_set_language()}; if not set, "en". If url given, this can be left empty.}

\item{cache}{Defaults to NULL. If given, it should be given either TRUE or FALSE. Typically set with \code{tw_enable_cache()} or \code{tw_disable_cache()}.}

\item{overwrite_cache}{Logical, defaults to FALSE. If TRUE, it overwrites the table in the local sqlite database. Useful if the original Wikidata object has been updated.}

\item{cache_connection}{Defaults to NULL. If NULL, and caching is enabled, \code{tidywikidatar} will use a local sqlite database. A custom connection to other databases can be given (see vignette \code{caching} for details).}

\item{disconnect_db}{Defaults to TRUE. If FALSE, leaves the connection to cache open.}
}
\value{
A character vector of length one with a Wikidata Q identifier.
}
\description{
Gets the Wikidata id of a Wikipedia page
}
\examples{
if (interactive()) {
  tw_get_qid_of_wikipedia_page_single(title = "Margaret Mead", language = "en")
}
}
