% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_get_id_from_wikipedia.R
\name{tw_get_wikipedia_base_api_url}
\alias{tw_get_wikipedia_base_api_url}
\title{Facilitates the creation of MediaWiki API base URLs}
\usage{
tw_get_wikipedia_base_api_url(url = NULL, title = NULL, language = NULL)
}
\arguments{
\item{url}{Full URL to a Wikipedia page. If given, title and language can be left empty.}

\item{title}{Title of a Wikipedia page or final parts of its url. If given, url can be left empty, but language must be provided.}

\item{language}{Two-letter language code used to define the Wikipedia version to use. If url given, this can be left empty.}
}
\value{
A charachter vector of base urls to be used with the MediaWiki API
}
\description{
Mostly used internally
}
\examples{

tw_get_wikipedia_base_api_url(title = "Margaret Mead", language = "en")
}
