% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_prune.R
\name{tw_prune}
\alias{tw_prune}
\title{Prune unnecessary item properties from the local cache}
\usage{
tw_prune(
  p_to_keep = NULL,
  p_to_remove = NULL,
  language = tidywikidatar::tw_get_language(),
  cache = NULL,
  cache_connection = NULL,
  disconnect_db = TRUE
)
}
\arguments{
\item{p_to_keep}{A character vector for properties in the format, e.g., \code{c("P123", "P124")}. Only these properties will be kept in the cache.}

\item{p_to_remove}{A character vector for properties in the format, e.g., \code{c("P123", "P124")}. These properties will be removed from cache.}

\item{language}{Defaults to language set with \code{tw_set_language()}; if not set, "en". Use "all_available" to keep all languages. For available language values, see https://www.wikidata.org/wiki/Help:Wikimedia_language_codes/lists/all}

\item{cache}{Defaults to NULL. If given, it should be given either TRUE or FALSE. Typically set with \code{tw_enable_cache()} or \code{tw_disable_cache()}.}

\item{cache_connection}{Defaults to NULL. If NULL, and caching is enabled, \code{tidywikidatar} will use a local sqlite database. A custom connection to other databases can be given (see vignette \code{caching} for details).}

\item{disconnect_db}{Defaults to TRUE. If FALSE, leaves the connection to cache open.}
}
\value{
Nothing, used for its side effects.
}
\description{
Prune unnecessary item properties from the local cache
}
\examples{

p_to_keep <- c("P31", "P21", "P569", "P570", "P19", "P20")

}
