% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tw_qid_airports}
\alias{tw_qid_airports}
\title{The Wikidata Q identifier of European airports found in Eurostat's \code{avia_par_} dataset}
\format{
A data frame with 429 rows and 1 column:
\describe{
\item{id}{Q identifiers}
}
}
\source{
\url{https://www.wikidata.org/wiki/Wikidata:Main_Page}
}
\usage{
tw_qid_airports
}
\description{
The Wikidata Q identifier of European airports found in Eurostat's \code{avia_par_} dataset
}
\keyword{datasets}
