% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tw_qid_meps}
\alias{tw_qid_meps}
\title{The Wikidata Q identifier of all members of the European Parliament since its establishment}
\format{
A data frame with 4581 rows and 1 column:
\describe{
\item{id}{Q identifiers}
}
}
\source{
\url{https://www.wikidata.org/wiki/Wikidata:Main_Page}
}
\usage{
tw_qid_meps
}
\description{
A dataset with all the Wikidata items that have "Q27169" (member of the European Parliament) for the property "P39" (position held).
}
\keyword{datasets}
