% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_get.R
\name{tw_search}
\alias{tw_search}
\title{Search string in Wikidata and return Wikidata id, label, and description.}
\usage{
tw_search(search, language = "en", limit = 10, wait = 1, cache = TRUE)
}
\arguments{
\item{search}{A string to be searched in Wikidata}

\item{language}{Language to be used for the search}

\item{limit}{Maximum numbers of responses to be given.}

\item{wait}{In seconds, defaults to 1. Time to wait between queries to Wikidata. If data are cached locally, wait time is not applied.}

\item{cache}{Logical, defaults to TRUE. If TRUE, search queries are stored in a local sqlite database located in the \code{wiki_search_db} folder within the local cache folder.}
}
\value{

}
\description{
Search string in Wikidata and return Wikidata id, label, and description.
}
\examples{

\dontrun{
tw_search(search = "Mihai Eminescu", language = "en")
tw_search(search = "Mihai Eminescu", language = "ro")
}

}
