% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_language_settings.R
\name{tw_set_language}
\alias{tw_set_language}
\alias{tw_get_language}
\title{Set language to be used by all functions}
\usage{
tw_set_language(language = NULL)

tw_get_language(language = NULL)
}
\arguments{
\item{language}{A character vector of length one, with a string of two letters such as "en". For a full list of available values, see: https://www.wikidata.org/wiki/Help:Wikimedia_language_codes/lists/all}
}
\value{
A two letter code for the language, if previously set; the same language as given to the function; or the default, \code{en} is none is given.
}
\description{
Defaults to "en".
}
\examples{
\donttest{
if (interactive()) {
  tw_set_language(language = "en")
}
}

tw_get_language()
}
