% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_caching.R
\name{tw_check_cache}
\alias{tw_check_cache}
\title{Check caching status in the current session, and override it upon request}
\usage{
tw_check_cache(cache = NULL)
}
\arguments{
\item{cache}{Defaults to NULL. If NULL, checks current cache settings. If given, returns given value, ignoring cache.}
}
\value{
Either TRUE or FALSE, depending on current cache settings.
}
\description{
Mostly used internally in functions, exported for reference.
}
\examples{
\donttest{
if (interactive()) {
  tw_check_cache()
}
}
}
