% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_check_qid.R
\name{tw_check_qid}
\alias{tw_check_qid}
\title{Ensures that input appears to be a valid Wikidata id}
\usage{
tw_check_qid(id)
}
\arguments{
\item{id}{A character vector of one or more Wikidata id.}
}
\value{
A character vector with only strings appearing to be Wikidata identifiers; possibly shorter than input
}
\description{
Mostly used internally by other functions.
}
\examples{

tw_check_qid(id = c("Q180099","q228822","Not an id","00180099", NA))
}
