% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_filter.R
\name{tw_filter}
\alias{tw_filter}
\title{Filter search result and keep only items with matching property and Q id}
\usage{
tw_filter(
  search,
  p,
  q,
  language = tidywikidatar::tw_get_language(),
  limit = 10,
  include_search = FALSE,
  wait = 0,
  cache = NULL,
  overwrite_cache = FALSE,
  cache_connection = NULL,
  disconnect_db = TRUE
)
}
\arguments{
\item{search}{A data frame generated by \code{tw_search()}, or a search query. If a data frame is given, language and limits are ignore.}

\item{p}{A character vector of length 1, a property. Must always start with the capital letter "P", e.g. "P31" for "instance of".}

\item{q}{A character vector of length 1, a wikidata id. Must always start with the capital letter "Q", e.g. "Q5" for "human being".}

\item{language}{Language to be used for the search. Can be set once per session with \code{tw_set_language()}. If not set, defaults to "en". For a full list, see https://www.wikidata.org/wiki/Help:Wikimedia_language_codes/lists/all}

\item{limit}{Maximum numbers of responses to be given.}

\item{include_search}{Logical, defaults to FALSE. If TRUE, the search is returned as an additional column.}

\item{wait}{In seconds, defaults to 0. Time to wait between queries to Wikidata. If data are cached locally, wait time is not applied. If you are running many queries systematically you may want to add some waiting time between queries.}

\item{cache}{Defaults to NULL. If given, it should be given either TRUE or FALSE. Typically set with \code{tw_enable_cache()} or \code{tw_disable_cache()}.}

\item{overwrite_cache}{Defaults to FALSE. If TRUE, overwrites cache.}

\item{cache_connection}{Defaults to NULL. If NULL, and caching is enabled, \code{tidywikidatar} will use a local sqlite database. A custom connection to other databases can be given (see vignette \code{caching} for details).}

\item{disconnect_db}{Defaults to TRUE. If FALSE, leaves the connection to cache open.}
}
\value{
A data frame with three columns, \code{id}, \code{label}, and \code{description}, filtered by the above criteria.
}
\description{
Filter search result and keep only items with matching property and Q id
}
\examples{
tw_search(search = "Margaret Mead", limit = 3) \%>\%
  tw_filter(p = "P31", q = "Q5")
}
