% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_get.R
\name{tw_get}
\alias{tw_get}
\title{Return (almost) all information from a Wikidata item in a tidy format}
\usage{
tw_get(id, cache = TRUE, overwrite_cache = FALSE)
}
\arguments{
\item{id}{A characther vector, must start with Q, e.g. "Q254" for Wolfgang Amadeus Mozart. Can also be a data frame of one row, typically generated with \code{tw_search()} or a combination of \code{tw_search()} and \code{tw_filter_first()}.}

\item{cache}{Logical, defaults to TRUE. If TRUE, it stores all retrieved data in a local sqlite database.}

\item{overwrite_cache}{Logical, defaultso to FALSE. If TRUE, it overwrites the table in the local sqlite database. Useful if the original Wikidata object has been updated.}
}
\value{
A data.frame (a tibble) with two columns: property and value
}
\description{
Return (almost) all information from a Wikidata item in a tidy format
}
\examples{
\dontrun{
tw_get("Q254")
}

}
