% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_get_cached_item.R
\name{tw_get_cache_file}
\alias{tw_get_cache_file}
\title{Gets location of cache file}
\usage{
tw_get_cache_file(type = "item", language = "all_available")
}
\arguments{
\item{type}{Defaults to "item". Type of cache file to output. Values typically used by \code{tidywikidatar} include "item", "search", and "qualifier".}

\item{language}{Defaults to "all_available". Use to limit the data to be cached. For available values, see https://www.wikidata.org/wiki/Help:Wikimedia_language_codes/lists/all}
}
\value{
A character vector of length one with location of item cache file.
}
\description{
Gets location of cache file
}
\examples{

tw_set_cache_folder(path = tempdir())
sqlite_cache_file_location <- tw_get_cache_file() # outputs location of cache file
}
