% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_get_property_with_details.R
\name{tw_get_property_with_details_single}
\alias{tw_get_property_with_details_single}
\title{Gets all details of a property}
\usage{
tw_get_property_with_details_single(id, p)
}
\arguments{
\item{id}{A characther vector, must start with Q, e.g. "Q254" for Wolfgang Amadeus Mozart.}

\item{p}{A character vector, a property. Must always start with the capital letter "P", e.g. "P31" for "instance of".}
}
\value{
A tibble, corresponding to the details for the given property. NULL if no relevant property found.
}
\description{
Gets all details of a property
}
\examples{
# Get "female form of label", including language
tw_get_property_with_details(id = "Q64733534", p = "P2521")
}
