% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_get.R
\name{tw_get_wikipedia}
\alias{tw_get_wikipedia}
\title{Get Wikidata description in given language}
\usage{
tw_get_wikipedia(id, language = "en", cache = TRUE)
}
\arguments{
\item{id}{A characther vector, must start with Q, e.g. "Q254" for Wolfgang Amadeus Mozart}

\item{language}{A character vector of length one, defaults to "en", must correspond to a two-letter language code.}

\item{cache}{Logical, defaults to TRUE. If TRUE, it stores all retrieved data in a local sqlite database.}
}
\value{
A charachter vector of length 1, with the Wikidata label in the requested languae.
}
\description{
Get Wikidata description in given language
}
\examples{


}
