% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_get.R
\name{tw_search}
\alias{tw_search}
\title{Search string in Wikidata and return Wikidata id, label, and description.}
\usage{
tw_search(search, language = "en", limit = 10, wait = 0.1, cache = NULL)
}
\arguments{
\item{search}{A string to be searched in Wikidata}

\item{language}{Language to be used for the search}

\item{limit}{Maximum numbers of responses to be given.}

\item{wait}{In seconds, defaults to 0.1. Time to wait between queries to Wikidata. If data are cached locally, wait time is not applied.}

\item{cache}{Defaults to NULL. If given, it should be given either TRUE or FALSE. Typically set with \code{tw_enable_cache()} or \code{tw_disable_cache()}.}
}
\value{
A data frame (a tibble) with three columns (id, label, and description), and as many rows as there are results (by default, limited to 10).
}
\description{
Search string in Wikidata and return Wikidata id, label, and description.
}
\examples{

\dontrun{
tw_search(search = "Mihai Eminescu", language = "en")
tw_search(search = "Mihai Eminescu", language = "ro")
}

}
